requirements.txt
django >= 1.9.5
djangorestframework >= 3.3.3
djangorestframework-gis >= 0.10
markdown >= 2.6.6
django-filter >= 0.13.0
celery >= 3.1
flower >= 0.9.1
psycopg2 >= 2.6
pinax-theme-bootstrap >= 7.9.0
django-user-accounts >= 1.3.1
pinax-webanalytics >= 2.0.1
django-constance[database] >= 1.2
django-forms-bootstrap == 3.0.1
django-crispy-forms >= 1.5.2
django-multiselectfield >= 0.1.3
pyproj >= 1.9.5

django-niamoto-data>=0.1
django-niamoto-plantnote>=0.1
django-niamoto-geoserver-admin>=0.1
django-niamoto-forest-digitizing>=0.1
